requirements.txt
PyPDF2
sentence-transformers
faiss-cpu
requests
python-dotenv